Fix for file upload issue (#45)
* Fixed backwards logic

If the 'noform' query parameter does not exist, then FormFile must be
used because the upload will be a multipart form.  When 'noform' is set,
the uploaded file is plain text in the body of the request and can be
simply copied out to the local file.  Prior to this commit, if 'noform'
was absent from the query, then the multipart form was copied (in
encoded form) to the file.

* Fixes PufferPanel/pufferd#44

This commit removes the dependency on the client to specify whether or
not the file upload is formatted as a multipart message or simply
included in the body of the request.  The condition is now determined
based on the Content-Type tag in the request header.